Store just the filename of downloaded media in a separate key
So that it can be simply found on the disk without needing to have
any extra logic in the display template. Then anyone trying to view
linked media has a choice of getting (or refetching if copy in S3 is
lost or damaged) the file from Twitter's servers or use a downloaded
version from S3.

Had to pass the tweet_id into the media_entities function to construct
the on-disk filename as it requires a "{tweet_id}-" prefix